requirements.txt
boto3>=1.26.0
google-cloud-billing>=1.7.0
google-api-python-client>=2.50.0
azure-identity>=1.12.0
azure-mgmt-compute>=29.0.0
requests>=2.28.0
pyscf>=2.2.0